requirements2.txt
pyspark==3.4.3
delta-spark==2.4.0
numpy==1.24.0
pandas==1.5.3
openpyxl==3.1.2
googlemaps==4.10.0
kafka-python==2.0.2
requests==2.31.0